rename: Model내에 있는 파일 이름 변경
