+++ src/Models/lottoListIssueUtils.js
